21540: Start a transaction in update_priorities
It is harmless if we are already in a transaction.

It's possible that not being in a transaction could be a root cause of
the deadlocks.  If it is not in a transaction when it does the ordered
locking, then the locks would be released immediately, and the update
priority update query would re-take the locks in arbitrary order.

Arvados-DCO-1.1-Signed-off-by: Peter Amstutz <peter.amstutz@curii.com>