22317: Replace compute image base.sh with an Ansible playbook
The hope is that this will be more maintainable and extendable, and
allow us to start building towards a general Ansible installer.

This Ansible playbook was written and tested against Ansible 8.7.0,
because Ansible 8.x seems to be the version that supports most of the
Pythons we care about. This might need refinement in the future.

The Ansible playbook supports all the same deployment configurations as
base.sh, but there are some differences in the way it goes about things:

* It no longer configures GRUB for cgroups v1 because Arvados supports
  cgroups v2 now.

* It installs software to /opt instead of /var/lib/arvados, because
  software doesn't belong in /var and none of this is Arvados software
  specifically looking for that path.

* It only installs either Docker or Singularity as required, not both.

* It builds Singularity in a temporary directory that it cleans up
  afterwards to save space in the AMI.

* It sets up the encrypted partition script as a systemd service rather
  than a cloud-init script so we can be more explicit about the ordering
  requirements. The service orders itself before SSH so it has a chance to
  finish before the Crunch dispatcher is able to SSH in. If it fails, that
  will be reflected in systemctl's system state, and therefore our default
  BootProbeCommand as well.

Arvados-DCO-1.1-Signed-off-by: Brett Smith <brett.smith@curii.com>