22349: Run RailsAPI as a standalone Passenger service
The primary motivation for this is to be able to run Passenger with the
same version of Ruby that RailsAPI depends on. On RHEL8, the official
Passenger packages depend on the system Ruby 2.5, which is too old for
us to use.

A secondary benefit is that this simplifies installation for everyone by
eliminating the need for a separate Passenger install and nginx
integration.

A tertiary benefit is that the systemd service definition can better
handle some preparation work that we were previously doing in the
postinst script.

Arvados-DCO-1.1-Signed-off-by: Brett Smith <brett.smith@curii.com>