22406: Revamp login_ldap_docker_test
I swear this tests all the same functionality, I just completely
reorganized it in order to get in running in a wider variety of
environments. Here are the key improvements:

The test suite now creates a Docker network, and runs all its containers
inside that. This setup requires less cooperation from the host
network. It also means we can hardcode names and ports of different
services inside the container network. "All" we need to do is create
proxies on the network gateway for the controller to contact test
services running on the host. And even that's cleaner because now we can
bind on that single network gateway interface instead of every interface.

The tests now launch containers that correspond to the distro they're
running on to ensure that the built controller will actually work.

This revealed that libpam-ldap has been removed from Debian 12 for lack
of maintenance. I have reorganized the test to create plain passwd
accounts on container start, and then authenticate against that. While
this changes the PAM backend, we still test all the same PAM
authentication functionality.

To help make all this happen, write more of the test in Go rather than
shell. This let me better organize different parts of the test, and
write better test assertions. I hope this will make it easier for the
next developer to follow and work on.

Arvados-DCO-1.1-Signed-off-by: Brett Smith <brett.smith@curii.com>