22433: Flatten RailsAPI postinst script
The previous version had most of the work in a giant `configure_version`
function and then just did a little status reporting at the end. Reverse
that relationship: have the main functionality be the top level of the
script, and move incomplete install reporting to a dedicated function.

Arvados-DCO-1.1-Signed-off-by: Brett Smith <brett.smith@curii.com>