22433: Report a failure status if RailsAPI postinst does not complete
In the past, we specifically wanted to return exit code 0 if RailsAPI
couldn't be installed because of missing configuration or supporting
database. This meant the message we reported about it was the last
thing users saw, and they could get to work fixing it.

However, we are really pushing to get people to use higher-level
installers like Salt and Ansible. In this scenario, the exit code 0 is a
problem: it means the installer will assume RailsAPI is good to go even
if it isn't.

Now we prioritize this case over the by-hand install and exit with an
error code if postinst cannot complete.

Arvados-DCO-1.1-Signed-off-by: Brett Smith <brett.smith@curii.com>