21547: Don't retry row_lock_for_priority_update deadlocks
It seems like retrying a deadlock is counter productive -- it ties up
a request slot and holds a transaction open.  I believe returning a
500 error and having the client retry will be more robust.

This also adjusts the locking query to select and order by the primary
key (id).  This is speculative and based on a stack overflow comment
suggesting that the doing lock ordering on a field other than the
primary key causes rows to be first acquired with share locks (out of
order) and then subsequently converted to exclusive locks (in order),
and this can be avoided by using the primary key.

Since I haven't been able to determine any other reason the ordered
locking, which is specifically implemented to prevent deadlocks but
deadlocks anyway, it's worth a shot.

Arvados-DCO-1.1-Signed-off-by: Peter Amstutz <peter.amstutz@curii.com>