22406: Remove hardcoding of arvados_test database name
Users should be allowed to use a different database if they
want. Testing that this actually works is outside the scope of this
branch, but we would like to remove obstacles to it happening, rather
than adding more.

zzzzz is assumed to always be a cluster running automated tests. If you
want a development cluster in your config.yml using a development
database, you need to give it a different cluster ID.

Arvados-DCO-1.1-Signed-off-by: Brett Smith <brett.smith@curii.com>