22320: Fix Repack bug causing segment replacements to be skipped.
repackTree must call replaceSegments with the newly computed
replacements -- not the merged set of new replacements plus
previously-applied replacements.  If it passes the merged set to
replaceSegments, some of the new replacements will be skipped.

For example, if a previous repack applied A->B, and the current repack
applies B->D and C->D, then the merged set includes A->D, and A is no
longer referenced, so replaceSegments will skip any->D.

Arvados-DCO-1.1-Signed-off-by: Tom Clegg <tom@curii.com>