22563: Remove locale configuration from build-compute-image
Arvados tests need this because they rely on having the en_US.UTF-8
locale to pass. Compute nodes do not need it. An administrator might
want to set up locales for their own reasons, but it's not necessary to
build a compute node, so it should be done elsewhere.

Arvados-DCO-1.1-Signed-off-by: Brett Smith <brett.smith@curii.com>