22602: Update arv-copy docs a bit
I switched some UUID references from "jutro" to "pirca".  I think the
original reason to use jutro is because pirca was supposed to be
cleaned up periodically (which hasn't happened) and they are
federated.  However, because we are not talking about federation in
this section, this is mostly just confusing.

There are still references to jutro under "copy a workflow" and "copy
a project" and I'm leaving them alone for now.

Arvados-DCO-1.1-Signed-off-by: Peter Amstutz <peter.amstutz@curii.com>