22608: Guard against no uuid param in find_object_by_uuid
See comments for rationale. The old code causes bad query construction
and crashes in Rails 7.2 because of
  <https://github.com/rails/rails/issues/52643>

While we can't upgrade to Rails 7.2 right now, the code is safe for
older versions. Since I spent the time to track this down, we might as
well make the change now and make a future upgrade easier.

Arvados-DCO-1.1-Signed-off-by: Brett Smith <brett.smith@curii.com>