22395: Use toggleResourceTrashed everywhere. Also detect isTrashed using items.
Technically we probably shouldn't allow a mixed group if isTrashed like before,
and with batching, it seems better to present an error for double trash/untrash
instead of performing multiple operations at once. In the future maybe we can
better tie isTrashed to the logic in the ToggleTrashAction component.

Arvados-DCO-1.1-Signed-off-by: Stephen Smith <stephen@curii.com>