22581: Added a "initial_path" field to published_ports
This is a relative path (and fragment, and query parameters) that
should be added to the URL when constructing the clickable link in
Workbench.

I thought of this as I was working on branch for the client side
support.  For a given container service we might want to start at a
page somewhere other than the root, or we might want to provide a
query parameter to the service.  This gives a way to do that.

Arvados-DCO-1.1-Signed-off-by: Peter Amstutz <peter.amstutz@curii.com>