null values fall back to zero when computing read_length
found refs #21074, original bugfix refs #22232

Null is not the same as zero, in SQL (null + a number) is null, not
the number.  As a result a single null value would cause the entire
row to be considered null (and Ruby would assign it zero size)
regardless of the size of the other columns in the row.  Use coalese()
to get the intended behavior.

Arvados-DCO-1.1-Signed-off-by: Peter Amstutz <peter.amstutz@curii.com>