Filter the development- prefix before sorting, refs #22767
Unfortunately 'development' sorts before numbers, so the
development-3.0.0 tag was continuing to show up at the top (resulting
in it assigning 3.1.0~dev versions even after 3.1.0 had been released
and tagged) when it should have been picking up 3.1.0 as the most
recent version.  Tweak the logic to get rid of the development prefix
before sorting.  With this, it correctly generates the versions as
3.2.0~dev.

Arvados-DCO-1.1-Signed-off-by: Peter Amstutz <peter.amstutz@curii.com>