22822: Fix VIRTUAL_ENV path in Python packages
The code we have to update the `activate` script is failing on at least
some distributions, so activating the virtualenv appears to work but is
ineffective.

Trying to modify shell source on the fly is inherently a dicey
proposition. Instead of doing that, let's rely on the fact that builds
happen in an ephemeral Docker container, and just build the "source"
virtualenv in the same path where the package will install it,
eliminating the need for any rewriting.

Arvados-DCO-1.1-Signed-off-by: Brett Smith <brett.smith@curii.com>