22680: Make 'any_searchable_columns' which excludes datetime columns
Can't use datetime in all-column search because Postgresql will try to
parse the string literal into a datetime during query compile time and
throw an error.  We'd prefer the comparison just evaluated to 'false'
but that's not what it does.  Stack Overflow suggests there a solution
involving trapping the error with a SQL stored procedure but the more
immediate fix is to just ignore the columns entirely.

The fact that we haven't seen this error before means no one is
relying on this behavior, while fixing it allows us to have passing
tests that check for a concrete security concern in the credentials
API.

Arvados-DCO-1.1-Signed-off-by: Peter Amstutz <peter.amstutz@curii.com>