22957: Build package test Docker images with Ansible
This is not a huge win by itself, but it means we can manage these
images with a single playbook rather than one Dockerfile per image. The
bigger win will come when we do the same for package build Dockerfiles,
which can reuse a lot of the code we've already written in
`install-test-env.yml`.

Arvados-DCO-1.1-Signed-off-by: Brett Smith <brett.smith@curii.com>