22958: Introduce distro_packages role
The point of this role is to abstract away differences between
distributions. This is the one place that knows about packaging
differences across distributions and takes care of it, so we don't have
to keep re-encoding the same logic across our playbooks.

Arvados-DCO-1.1-Signed-off-by: Brett Smith <brett.smith@curii.com>