22958: Update Python packages to depend on libcurl4
This is sort of a side effect of the switch to building Docker images
with Ansible, but I wanted it to be a separate commit for
clarity. `install-dev-tools.yml` installs `libcurl4-openssl-dev` because
it's clearly the preferred version: it depends on libcurl4, which in
turn is the version used by curl. The package build Dockerfiles use
`libcurl4-gnutls-dev` which depends on `libcurl3-gnutls` which is
clearly second-rate.

Arvados-DCO-1.1-Signed-off-by: Brett Smith <brett.smith@curii.com>