22866: Document single-node install with Ansible
This is tested and we expect this will be easier for most people to
deploy.

Part of what's changing here is the documentation page does not have a
bunch of forks for different options. Instead it prioritizes getting the
simplest possible install working with minimum fuss. I think this is
what most people installing with this option want. The provided
configuration templates give pointers for ways you might change the
configuration in more advanced ways.

Arvados-DCO-1.1-Signed-off-by: Brett Smith <brett.smith@curii.com>