20311: Generate _version.py from a setuptools subcommand
This restructures arvados_version.py so that metadata generation happens
in a dedicated class (useful for other build tooling like
build_docker_image.py) and then it provides a setuptools subcommand to
write _version.py where needed. This provides cleaner separation of
concerns and gives us a better base to build more tooling from.

As part of this change, we take care to no longer install tests as part
of the wheel we build. We shouldn't have been doing this anyway, and now
it's required for the tests to run as expected.

Arvados-DCO-1.1-Signed-off-by: Brett Smith <brett.smith@curii.com>