Restart Docker after updating daemon.json
This was not in the original version of the code because it was written
to build cloud compute images where we just needed to write the right
configuration. Now that it supports single-node installs and maybe more
in the future, we need to make sure services get new state. No issue #.

Arvados-DCO-1.1-Signed-off-by: Brett Smith <brett.smith@curii.com>