23115: Build Workbench Docker image with Ansible
As part of a general code deduplication effort. In particular, this
unblocks the removal of `arvados-server install`.

Since I had to make significant changes to the integration test script
to match, I tried to make it match the Makefile better, using more of
the same variable names and using those variables for configuration.

Instead of copying/mounting the Arvados source as part of the image
build, instead we build dependencies from source into a dedicated
volume. There are pros and cons to this approach. One pro is that since
this stuff tends to churn more often, you can wipe the volume and
rebuild those dependencies without needing to rebuild the whole image.

Arvados-DCO-1.1-Signed-off-by: Brett Smith <brett.smith@curii.com>