23191: Document cluster activity workflow and inputs
With this change, I decided to remove support for the --days
option. Given that --end defaults to today, --days and --start achieve
the same thing. I decided it was better to just support one than
document that they're mutually exclusive. There's a case to be made for
either one, but I decided to prefer the one that doesn't change meaning
over time.

Arvados-DCO-1.1-Signed-off-by: Brett Smith <brett.smith@curii.com>