23213: Memoize property validation to avoid infinite re-renders
From what I gather, since redux-form v8, if the validation array changes,
redux-form will re-validate and re-render. Since this validation includes an
anonymous function (generated by matchTagValues), multiple runs of
getValidation will yield the "same" but not equal functions - as they will be
multiple instances of the same function. When the function is not defined
once like in validators.tsx but generated on the fly, we need to memoize it so
that it will re-use the same instance of the validator func.

Arvados-DCO-1.1-Signed-off-by: Stephen Smith <stephen@curii.com>