23136: Remove locking from FUSE event handler
As best we can tell today, you should *not* hold the FUSE lock when
invalidating inodes and entries. This is consistent with both the
libfuse reference documentation and included examples:

https://libfuse.github.io/doxygen/fuse__lowlevel_8h.html#a9cb974af9745294ff446d11cba2422f1
https://libfuse.github.io/doxygen/notify__inval__inode_8c.html

`new_entry` and `del_entry` do eventually lead to calls that may require
the lock. However, these functions directly just add the request to a
queue that is processed in a separate thread that acquires the lock when
needed.

If there's no need to acquire the FUSE lock here, then there's no need
to fuss with the collection lock either. This code doesn't change
anything about the underlying collection; it's only propagating changes
up to the higher levels of the FUSE stack.

So: remove all the lock manipulation going on here.

Similarly, it seems like you should *not* call `invalidate_entry` for an
entry that still exists but has merely had its contents
changed. Invalidating the inode should be sufficient.

Arvados-DCO-1.1-Signed-off-by: Brett Smith <brett.smith@curii.com>