23227: Evaluate WorkReuse.enableReuse requirement field as expression.
The CWL v1.1 specification says that "enableReuse" field under the
"WorkReuse" requirement can be a boolean or an expression. For container
requests associated with a CWL command-line tool, evaluating the
expression is now implemented.

Notice that for the requirement "arv:ReuseRequirement" as part of the
Arvados extension for CWL v1.0, the "enableReuse" field is defined as a
boolean. Therefore, no evaluation takes place.

When both "WorkReuse" and "arv:ReuseRequirement" requirements are
specified, the standardized "WorkReuse" takes precedence.

This issue is related to #23214.

Arvados-DCO-1.1-Signed-off-by: Zoë Ma <zoe.ma@curii.com>