23235: Fix memory leak.
Balancer was not garbage collected after completion because the memory
reporting goroutine had a reference to it, and was waiting forever on
a stopped ticker channel instead of exiting.

Also, adjust timer loop to always show memory stats at the end of a
balancing run.

Arvados-DCO-1.1-Signed-off-by: Tom Clegg <tom@curii.com>