23235: Fix unneeded collection updates when desired_replication==0.
Previous code was doing no-op updates because the "does
storage_classes_confirmed already have the correct value" code was
incorrect when desired_replication == 0.  (The correct value is [],
not storage_classes_desired.)

Also skip a json.Marshal if the result won't be needed because we're
not updating.

Arvados-DCO-1.1-Signed-off-by: Tom Clegg <tom@curii.com>