23293: Do not consume the -h argument passed to subcommand.
If the script is called in the following manner --

  arvcli.py subcommand -h

"-h" is not immediately consumed by the argument parser (and causing a
SystemExit). Instead, it is left to the subcommand subparser to handle
the request for help.

With this, the pass-through commands' own, existing command-line parsing
code can be left in place.

Arvados-DCO-1.1-Signed-off-by: Zoë Ma <zoe.ma@curii.com>