23304: Update Docker API version to 1.48
We need to upgrade to a minimum 1.44 to support Docker 29. We've been
using Docker 28 for a while, so just use its API version.

The `VirtualSize` field of images was deprecated in API 1.43 and removed
in API 1.44. The `Size` field is documented to have the same information.

Arvados-DCO-1.1-Signed-off-by: Brett Smith <brett.smith@curii.com>